# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tokens": {
      "post": {
        "operationId": "ApiTokensController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token created successfully"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Create new API token",
        "tags": [
          "API Tokens Management"
        ]
      },
      "get": {
        "operationId": "ApiTokensController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tokens fetched successfully"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get all API tokens",
        "tags": [
          "API Tokens Management"
        ]
      }
    },
    "/api/v1/tokens/{uuid}": {
      "delete": {
        "operationId": "ApiTokensController_delete",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "description": "UUID of the API token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token deleted successfully"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Delete API token",
        "tags": [
          "API Tokens Management"
        ]
      }
    },
...
    "/api/v1/nodes/create": {
      "post": {
        "description": "Create a new node",
        "operationId": "NodesController_createNode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Node created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateNodeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Create Node",
        "tags": [
          "Nodes Controller"
        ]
      }
    },
    "/api/v1/nodes/enable/{uuid}": {
      "get": {
        "description": "Enable node to further use",
        "operationId": "NodesController_enableNode",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "description": "Node UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Node enabled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnableNodeResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Enable Node",
        "tags": [
          "Nodes Controller"
        ]
      }
    },
    "/api/v1/nodes/restart/{uuid}": {
      "get": {
        "description": "Restart node",
        "operationId": "NodesController_restartNode",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "description": "Node UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Node restarted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RestartNodeResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Restart Node",
        "tags": [
          "Nodes Controller"
        ]
      }
    },
    "/api/v1/inbounds": {
      "get": {
        "description": "Get inbounds",
        "operationId": "InboundsController_getInbounds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get inbounds",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetInboundsResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get Inbounds",
        "tags": [
          "Inbounds Controller"
        ]
      }
    },
...
  },
  "info": {
    "title": "Remnawave API Schema",
    "description": "API for Remnawave",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "JWT obtained login."
      }
    },
    "schemas": {
...
      "CreateApiTokenRequestDto": {
        "type": "object",
        "properties": {
          "tokenName": {
            "type": "string"
          },
          "tokenDescription": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "tokenName",
          "tokenDescription"
        ]
      },
...
      "CreateNodeRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 5
          },
          "address": {
            "type": "string",
            "minLength": 2
          },
          "port": {
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false
          },
          "isBillTrackingActive": {
            "type": "boolean",
            "default": false
          },
          "trafficLimitBytes": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "notifyPercent": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 100,
            "exclusiveMaximum": false
          },
          "billDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "billCycle": {
            "description": "Bill cycle",
            "type": "string",
            "enum": [
              "MONTH",
              "MONTH",
              "YEAR"
            ],
            "nullable": true
          }
        },
        "required": [
          "name",
          "address",
          "billCycle"
        ]
      },
      "CreateNodeResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "properties": {
              "uuid": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "address": {
                "type": "string"
              },
              "port": {
                "type": "integer",
                "nullable": true
              },
              "isConnected": {
                "type": "boolean",
                "default": false
              },
              "isDisabled": {
                "type": "boolean",
                "default": false
              },
              "isConnecting": {
                "type": "boolean",
                "default": false
              },
              "isNodeOnline": {
                "type": "boolean",
                "default": false
              },
              "isXrayRunning": {
                "type": "boolean",
                "default": false
              },
              "lastStatusChange": {
                "nullable": true
              },
              "lastStatusMessage": {
                "type": "string",
                "nullable": true
              },
              "xrayVersion": {
                "type": "string",
                "nullable": true
              },
              "isBillTrackingActive": {
                "type": "boolean",
                "default": false
              },
              "billDate": {
                "nullable": true
              },
              "billCycle": {
                "type": "string",
                "enum": [
                  "MONTH",
                  "MONTH",
                  "YEAR"
                ],
                "nullable": true
              },
              "trafficLimitBytes": {
                "type": "integer",
                "nullable": true
              },
              "trafficUsedBytes": {
                "type": "integer",
                "nullable": true
              },
              "notifyPercent": {
                "type": "integer",
                "nullable": true
              },
              "createdAt": {},
              "updatedAt": {}
            },
            "required": [
              "uuid",
              "name",
              "address",
              "port",
              "lastStatusChange",
              "lastStatusMessage",
              "xrayVersion",
              "billDate",
              "billCycle",
              "trafficLimitBytes",
              "trafficUsedBytes",
              "notifyPercent",
              "createdAt",
              "updatedAt"
            ]
          }
        },
        "required": [
          "response"
        ]
      },
      "EnableNodeResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "properties": {
              "uuid": {
                "type": "string",
                "format": "uuid"
              },
              "isDisabled": {
                "type": "boolean",
                "default": false
              }
            },
            "required": [
              "uuid"
            ]
          }
        },
        "required": [
          "response"
        ]
      },
      "RestartNodeResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "properties": {
              "eventSent": {
                "type": "boolean"
              }
            },
            "required": [
              "eventSent"
            ]
          }
        },
        "required": [
          "response"
        ]
      },
      "GetInboundsResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "tag": {
                  "type": "string"
                },
                "uuid": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "required": [
                "tag",
                "uuid"
              ]
            }
          }
        },
        "required": [
          "response"
        ]
      },
...
    }
  }
}